+++ chem_mat_data/utils.py
@@ -33,6 +33,7 @@ TEMPLATE_PATH = os.path.join(PATH, 'templates')
 TEMPLATE_ENV = j2.Environment(loader=j2.FileSystemLoader(TEMPLATE_PATH))
 
 METADATA_PATH = os.path.join(PATH, 'metadata.yml')
+DOCS_PATH = os.path.join(os.path.dirname(PATH), 'docs')
 
 # MISC FUNCTIONS
 # ==============